ETL: Add support for reading/writing multiple files
This change adds support for reading and/or writing multiple files.

When reading RLP file(s), the -targetSize option can be used to
indicate that the output SSZ should be written to multiple output
files of (approximately) the targeted size.

When reading SSZ, if multiple files are passed on the command line,
they are all read in and written into one SSZ archive.